% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.rapidsplit}
\alias{loo.rapidsplit}
\title{Compute leave-one-out reliabilities}
\usage{
loo.rapidsplit(x)
}
\arguments{
\item{x}{A \code{rapidsplit} object.}
}
\value{
A \code{data.frame} with 3 columns: (1) \strong{subj} holds the subject ID (character),
(2) \strong{loo.r} holds the reliability if this subject is left out, and
(3) \strong{contribution} the full-sample reliability minus the leave-one-out reliability -
aids in interpretation since this value is positive when reliability is improved by the subject
being in the sample.
}
\description{
Per subject, compute what the reliability would be if they were left out.
}
\examples{
# Get a rapidsplit object to compute leave-one-out reliabilities from
data(foodAAT)
myrel <- rapidsplit.chunks(data=foodAAT,
                           subjvar="subjectid",
                           aggvar="RT",
                           splits=400,
                           split.chunksize=200,
                           sample.chunksize=50)

loo.rapidsplit(myrel)


}
\author{
Sercan Kahveci
}
