% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rapidsplithalf}
\alias{rapidsplithalf}
\alias{rapidsplithalf-package}
\title{rapidsplithalf package}
\description{
To learn more about rapidsplithalf, view the introductory vignette:
\code{vignette("rapidsplithalf",package="rapidsplithalf")}
}
