% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearmanbrown.R
\name{spearmanBrown}
\alias{spearmanBrown}
\title{Spearman-Brown correction
Perform a Spearman-Brown correction on the provided correlation score.}
\usage{
spearmanBrown(r, ntests = 2)
}
\arguments{
\item{r}{To-be-corrected correlation coefficient.}

\item{ntests}{An integer indicating how many times larger the full test is, 
for which the corrected correlation coefficient is being computed.}
}
\value{
Spearman-Brown corrected correlation coefficients.
}
\description{
Spearman-Brown correction
Perform a Spearman-Brown correction on the provided correlation score.
}
\details{
When \code{ntests=2}, the formula will compute what the correlation coefficient would be 
if the test were twice as long.
}
\examples{
spearmanBrown(.5)

}
